src/Entrega 2/Backend/Backend/auth-service/requirements.txt
fastapi>=0.110.0
pydantic>=2.6.1
pydantic-core>=2.16.2
uvicorn>=0.27.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.28
PyJWT>=2.8.0
email-validator>=2.2.0
cryptography>=42.0.5
requests>=2.32.3
httpx>=0.27.0
python-multipart>=0.0.9
psycopg2-binary
pymysql

